feat: Add Cache class to store data in Redis
The code changes include adding a new class called Cache, which allows storing data in Redis using a random key. This class initializes an instance of the Redis client and flushes the instance using flushdb. The store method stores the input data in Redis and returns the key. This addition enhances the functionality of the code by providing a convenient way to cache and retrieve data using Redis.