feat: Add Cache class to store data in Redis
This commit adds a new class called Cache to the codebase. The Cache class allows storing data in Redis using a random key. It initializes an instance of the Redis client and flushes the instance using flushdb. The store method is used to store the input data in Redis and returns the key. This addition enhances the functionality of the code by providing a convenient way to cache and retrieve data using Redis.